MODIFY: 카테고리마다 fetch 추가 성공
