+++ public/data/CategoryListData.json
@@ -1,66 +1,82 @@
 {
   "test": [
     {
-      "id": 1,
+      "id": 0,
       "img": "./images/fnvl.PNG",
       "name": "전체보기",
       "path": "main"
     },
     {
-      "id": 2,
+      "id": 1,
       "img": "./images/테크가전.jpg",
-      "name": "테크·가전"
+      "name": "테크·가전",
+      "path": "category"
+    },
+    {
+      "id": 2,
+      "img": "./images/디자인소품.jpg",
+      "name": "디자인소품",
       "path": "category"
     },
     {
       "id": 3,
-      "img": "./images/패션잡화.jpg",
-      "name": "패션·잡화"
+      "img": "./images/기부후원.jpg",
+      "name": "기부·후원",
+      "path": "category"
     },
     {
       "id": 4,
-      "img": "./images/뷰티.jpg",
-      "name": "뷰티"
+      "img": "./images/여행레저.jpg",
+      "name": "여행·레저",
+      "path": "category"
     },
     {
       "id": 5,
-      "img": "./images/푸드.jpg",
-      "name": "푸드"
+      "img": "./images/모임.jpg",
+      "name": "모임",
+      "path": "category"
     },
     {
       "id": 6,
-      "img": "./images/홈리빙.jpg",
-      "name": "홈리빙"
+      "img": "./images/푸드.jpg",
+      "name": "푸드",
+      "path": "category"
     },
     {
       "id": 7,
-      "img": "./images/디자인소품.jpg",
-      "name": "디자인소품"
+      "img": "./images/홈리빙.jpg",
+      "name": "홈리빙",
+      "path": "category"
     },
     {
       "id": 8,
-      "img": "./images/여행레저.jpg",
-      "name": "여행·레저"
+      "img": "./images/소셜캠페인.png",
+      "name": "소셜·캠페인",
+      "path": "category"
     },
     {
       "id": 9,
-      "img": "./images/스포츠모빌리티.jpg",
-      "name": "스포츠·모빌리티"
+      "img": "./images/패션잡화.jpg",
+      "name": "패션·잡화",
+      "path": "category"
     },
     {
       "id": 10,
       "img": "./images/반려동물.jpg",
-      "name": "반려동물"
+      "name": "반려동물",
+      "path": "category"
     },
     {
-      "id": 9,
+      "id": 11,
       "img": "./images/스포츠모빌리티.jpg",
-      "name": "스포츠·모빌리티"
+      "name": "스포츠·모빌리티",
+      "path": "category"
     },
     {
-      "id": 9,
-      "img": "./images/스포츠모빌리티.jpg",
-      "name": "스포츠·모빌리티"
+      "id": 12,
+      "img": "./images/교육키즈.jpg",
+      "name": "교육·키즈",
+      "path": "category"
     }
   ]
 }
