+++ src/Pages/Main/Components/CategoryList/Category.js
@@ -1,16 +1,8 @@
 import React, { Component } from 'react';
 
 class Category extends Component {
-  constructor() {
-    super();
-    this.state = {
-      cateName: '',
-    };
-  }
-
   render() {
     const { category, categoryClick } = this.props;
-    // console.log('여긴카테고리' + ca);
     return (
       <div
         className="categoryContent"
