+++ src/Pages/Main/Components/CategoryList/CategoryList.js
@@ -1,6 +1,5 @@
 import React, { Component } from 'react';
-import MainProductList from '../ProductList/MainProductList';
-import Category from './Category';
+// import Category from './Category';
 
 import './CategoryList.scss';
 
@@ -8,55 +7,55 @@ class CategoryList extends Component {
   constructor() {
     super();
     this.state = {
-      categoryData: [],
-      categoryId: 1,
+      // categoryData: [],
     };
   }
 
-  componentDidMount = () => {
-    this.categoryMount();
-  };
+  // componentDidMount = () => {
+  //   this.categoryMount();
+  // };
 
   //카테고리 mock data로 찍기
-  categoryMount = () => {
-    fetch('http://localhost:3000/data/categoryListData.json', {
-      method: 'GET',
-    })
-      .then(res => res.json())
-      .then(data => {
-        this.setState({
-          categoryData: data.test,
-        });
-      });
-  };
+  // categoryMount = () => {
+  //   fetch('http://localhost:3000/data/categoryListData.json', {
+  //     method: 'GET',
+  //   })
+  //     .then(res => res.json())
+  //     .then(data => {
+  //       this.setState({
+  //         categoryData: data.test,
+  //       });
+  //     });
+  // };
 
-  category = data => {
-    // console.log('과연 ? >>>>>' + data);
-    this.setState({
-      categoryId: data,
-    });
-  };
+  // category = data => {
+  //   // console.log('과연 ? >>>>>' + data);
+  //   this.setState({
+  //     categoryId: data,
+  //   });
+  // };
 
   render() {
-    const { categoryData, categoryId } = this.state;
-    // console.log('처음엔? >>>>' + this.state.categoryId);
-    // console.log('나중엔? >>>>' + this.state.categoryId);
+    const { categoryData } = this.props;
     return (
       <>
         <div className="categoryContainer">
           <div className="categoryList">
             {categoryData.map((data, index) => {
               return (
-                <Category
-                  category={data}
-                  key={index}
-                  categoryClick={this.category}
-                />
+                <div
+                  className="categoryContent"
+                  onClick={() => this.props.categoryClick(data)}
+                >
+                  <div className="categoryImg">
+                    <img src={data.img} alt={data.id} />
+                  </div>
+                  <span className="categoryName">{data.name}</span>
+                </div>
               );
             })}
           </div>
         </div>
-        <MainProductList id={categoryId} />
       </>
     );
   }
